# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "description": "API for managing users and their profiles",
        "contact": {
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a list of all users with their profiles",
                "operationId": "ec6e84f2bf746bf6f82f7aa1121e82a3",
                "responses": {
                    "200": {
                        "description": "A list of users with their profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserWithProfile"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user with basic profile information",
                "operationId": "1c9fad9a08059ec893d466dce9dc94d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithProfile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a specific user by ID with their profile",
                "operationId": "090d4825261951eb2fe868586769d886",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested user with their profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update an existing user with their profile information",
                "operationId": "110589f3ae608b1024bef1b0e8eeb66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithProfile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user and their associated profile",
                "operationId": "651f7723be5ebd2193026e079cbb8074",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User and their profile deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "messages": {
                                            "properties": {
                                                "success": {
                                                    "type": "string",
                                                    "example": "User and profile deleted successfully."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Profile": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "example": "https://example.com",
                        "nullable": true
                    },
                    "instagram": {
                        "type": "string",
                        "example": "johndoe_ig",
                        "nullable": true
                    },
                    "facebook": {
                        "type": "string",
                        "example": "https://facebook.com/johndoe",
                        "nullable": true
                    },
                    "linkedin": {
                        "type": "string",
                        "example": "https://linkedin.com/in/johndoe",
                        "nullable": true
                    },
                    "twitter_x": {
                        "type": "string",
                        "example": "@johndoe_tw",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "profile.jpg",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "example": "A short bio about John Doe.",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserWithProfile": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "properties": {
                            "profile": {
                                "$ref": "#/components/schemas/Profile"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "NewUserRequest": {
                "required": [
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "newuser"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "new@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "secure123"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "New",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "example": "User",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "updateduser"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "updated@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "newsecure"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Updated",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "example": "User",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "errors": {
                        "properties": {
                            "username": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The username field is required."
                                }
                            },
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The email field must be a valid email address."
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The password field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}